medical_classification_demo/requirements.txt
astroid==2.4.2
certifi==2020.6.20
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
h5py==2.10.0
imageio==2.9.0
isort==4.3.21
Keras==2.4.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.0
mccabe==0.6.1
networkx==2.4
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
pip==20.1.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scipy==1.5.2
seaborn==0.10.1
setuptools==49.2.0.post20200714
six==1.15.0
tifffile==2020.7.24
toml==0.10.1
typed-ast==1.4.1
wheel==0.34.2
wincertstore==0.2
wrapt==1.12.1
